refactor: :recycle: refactor folder structure and code
