+++ components/styled/index.js
@@ -1,0 +1,6 @@
+import Label from './label'
+import StyledLink from './link'
+
+export { StyledLink }
+export { Label }
+export { SuccessMessage, ErrorMessage } from './message'
