+++ components/svg/logo.js
