+++ pages/about.js
@@ -1,19 +1,16 @@
 import { Heading, Text } from '@chakra-ui/react'
-import Container from '../components/Container'
-import ContentWrapper from '../components/contentWrapper'
+import Container from '../layout/container'
 
 const About = () => (
-	<Container m='auto'>
-		<ContentWrapper>
-			<Heading as='h1' variant='h2'>
-				About 👨‍💻
-			</Heading>
-			<Text>
-				This page will contain information about my background, skillset and
-				hobbies outside of tech.
-			</Text>
-			<Text>It will also contain info about how to get in touch with me.</Text>
-		</ContentWrapper>
+	<Container title='About | Nitesh Seram'>
+		<Heading as='h1' variant='h2'>
+			About 👨‍💻
+		</Heading>
+		<Text>
+			This page will contain information about my background, skillset and
+			hobbies outside of tech.
+		</Text>
+		<Text>It will also contain info about how to get in touch with me.</Text>
 	</Container>
 )
