--- pages/api/hello.js
@@ -1,6 +1,0 @@
-// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-
-export default (req, res) => {
-	res.statusCode = 200
-	res.json({ name: 'John Doe' })
-}
