+++ pages/projects.js
@@ -1,22 +1,19 @@
 import { Heading, Text } from '@chakra-ui/react'
-import Container from '../components/Container'
-import ContentWrapper from '../components/contentWrapper'
+import Container from '../layout/container'
 
 const Projects = () => (
-	<Container>
-		<ContentWrapper>
-			<Heading as='h1' variant='h2'>
-				Projects 🎲
-			</Heading>
-			<Text>
-				This page will contain a list of projects I&apos;ve made or have worked
-				on.
-			</Text>
-			<Text>
-				Each project will contain a summary, live link (if applicable) and a
-				link to the project repository.
-			</Text>
-		</ContentWrapper>
+	<Container title='Projects | Nitesh Seram'>
+		<Heading as='h1' variant='h2'>
+			Projects 🎲
+		</Heading>
+		<Text>
+			This page will contain a list of projects I&apos;ve made or have worked
+			on.
+		</Text>
+		<Text>
+			Each project will contain a summary, live link (if applicable) and a link
+			to the project repository.
+		</Text>
 	</Container>
 )
